Support load time of thread safety tunning (#2)
The PR adds the option to configure (at load time) whether or not we want ijson to be thread safe (using mutex). This option can be set using `init_shared_string_cache` function and can only happened once at load time before any ijson value is created.

If thread safe is set, the string cache is wrap with mutex that allows it to be accessed from multiple threads.

Notice, the PR changes the type of `rc` to be AtomicUsize. A benchmark need to be done to make sure this change does not effect the use-cases where thread safety is not needed (promised by other means such as single threaded application).