Merge unsafe string (#5)
* add unsafe_string

* call ctor

* add init_cache

* add test for thread_unsafe

* rename thread unsafe test

* repalce feature thread_unsafe with thread_safe

* fix CI to run with no-default-features

* Implement defrag support for `IValue`. (#1)

The PR adds support for defrag on `IValue`. The defrag uses the provided `DefragAllocator` to reallocate all the memory of the `IValue`.

### Shared String Support

In general, it is considered easy to reallocate all the pointer of the `IValue` the only exception is the shared strings which we can not simply reallocate and free the old pointer because there might be more references that points to it.

To solve this issue we provide a way to clean the shared strings dictionary. This will cause new strings to be allocated from scratch and old shared strings will be freed when there will have no more references.

Notice that this means that now a shared string might not actually be shared between all jsons. Two json might have the same strings but with a different instance. This means that we can no longer count on pointer value for strings hash function, we must perform the hash function on the string value itself.

* Support load time of thread safety tunning (#2)

The PR adds the option to configure (at load time) whether or not we want ijson to be thread safe (using mutex). This option can be set using `init_shared_string_cache` function and can only happened once at load time before any ijson value is created.

If thread safe is set, the string cache is wrap with mutex that allows it to be accessed from multiple threads.

Notice, the PR changes the type of `rc` to be AtomicUsize. A benchmark need to be done to make sure this change does not effect the use-cases where thread safety is not needed (promised by other means such as single threaded application).

---------

Co-authored-by: Guy Korland <gkorland@gmail.com>
Co-authored-by: Meir Shpilraien (Spielrein) <meir@redis.com>